Ignore some files and package up the app
This project is now equipped with a Gemfile. You'll be able to install
all the necesary software to build and run this site just by running
`bundle` in the project directory. This, of course, assumes you have
Bundler and Ruby installed.

Additionally, I've added .gitignore so we can never worry about accidentally
committing the _site directory or any .DS_Store files.